corrected unittests mistake in chapter 5 and edited notebook titles in line with filenames
